# docs/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Vitta API
  description: |-
    API Specification for Vitta
  contact:
    email: prabhuomkar@pm.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://vitta-demo.fly.io
...
paths:
...
  /v1/payees:
    post:
      tags:
        - Payees
      summary: Create a new payee
      operationId: createPayee
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payee'
        required: true
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payee'
        '400':
          description: Bad request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Error
    get:
      tags:
        - Payees
      summary: Get all existing payees
      operationId: getPayees
      security:
        - basicAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payees'
        '401':
          description: Unauthenticated
        '500':
          description: Internal Error
  /v1/payees/{id}:
    patch:
      tags:
        - Payees
      parameters:
        - name: id
          in: path
          description: ID of Payee
          required: true
          schema:
            type: string
            format: uuid
      summary: Update an existing payee
      operationId: updatePayee
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payee'
        required: true
      responses:
        '204':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Error
    delete:
      tags:
        - Payees
      parameters:
        - name: id
          in: path
          description: ID of Payee
          required: true
          schema:
            type: string
            format: uuid
      summary: Delete an existing payee
      operationId: deletePayee
      security:
        - basicAuth: []
      responses:
        '204':
          description: Success
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Error
  /v1/transactions:
    post:
      tags:
        - Transactions
      summary: Create a new transaction
      operationId: createTransaction
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
        required: true
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Error
    put:
      tags:
        - Transactions
      summary: Import transactions
      operationId: importTransactions
      security:
        - basicAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: string
              format: binary
        required: true
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResult'
        '400':
          description: Bad request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Error
    get:
      tags:
        - Transactions
      summary: Get all existing transactions
      operationId: getTransactions
      security:
        - basicAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transactions'
        '401':
          description: Unauthenticated
        '500':
          description: Internal Error
  /v1/transactions/{id}:
    patch:
      tags:
        - Transactions
      parameters:
        - name: id
          in: path
          description: ID of Transaction
          required: true
          schema:
            type: string
            format: uuid
      summary: Update an existing transaction
      operationId: updateTransaction
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
        required: true
      responses:
        '204':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Error
    delete:
      tags:
        - Transactions
      parameters:
        - name: id
          in: path
          description: ID of Transaction
          required: true
          schema:
            type: string
            format: uuid
      summary: Delete an existing transaction
      operationId: deleteTransaction
      security:
        - basicAuth: []
      responses:
        '204':
          description: Success
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Error
components:
  schemas:
...
    Payee:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 01926b84-93cb-75b9-b445-a5a744f7a59f
          readOnly: true
        name:
          type: string
          example: Swiggy
        createdAt:
          type: string
          format: date-time
          example: 2024-10-07T01:02:03Z
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: 2024-10-07T01:02:03Z
          readOnly: true
    Payees:
      type: array
      items:
        $ref: '#/components/schemas/Payee'
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 01926b84-93cb-75b9-b445-a5a744f7a59f
          readOnly: true
        accountId:
          type: string
          format: uuid
          example: 01926b84-93cb-75b9-b445-a5a744f7a59f
        payeeId:
          type: string
          format: uuid
          example: 01926b84-93cb-75b9-b445-a5a744f7a59f
        categoryId:
          type: string
          format: uuid
          example: 01926b84-93cb-75b9-b445-a5a744f7a59f
        name:
          type: string
          example: House Party Food
        notes:
          type: string
          example: Sample markdown notes
        credit:
          type: number
          format: float
          multipleOf: 0.2
          example: 12.25
        debit:
          type: number
          format: float
          multipleOf: 0.2
          example: 10.49
        createdAt:
          type: string
          format: date-time
          example: 2024-10-07T01:02:03Z
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: 2024-10-07T01:02:03Z
          readOnly: true
    Transactions:
      type: array
      items:
        $ref: '#/components/schemas/Transaction'
    TransactionsResult:
      type: object
      properties:
        total:
          type: integer
          format: int64
          example: 10
        imported:
          type: integer
          format: int64
          example: 6
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic